+++ lib/storage/data/temperature/TemperatureLifetimeStats.h
@@ -1,11 +1,10 @@
 #pragma once
 
-#include <float.h>
 #include <Arduino.h>
 
 struct __attribute__((packed)) TemperatureLifetimeStats {
-    float minTemp = FLT_MAX;
-    float maxTemp = FLT_MIN;
+    int16_t minTemp = INT16_MAX;
+    int16_t maxTemp = INT16_MIN;
 
     uint32_t minDay : 5;
     uint32_t minMonth : 4;
